Removed global variables

+++ min-mon.py
@@ -422,7 +422,7 @@ def getEthminerData():
 
 	#	Prevent a divide by zero error
 	if ethShares > 0:
-		data['ethsharesperhour'] = "{0:.2f}".format(ethShares / ( ethUptimeMin / 60 ))
+		data['ethsharesperhour'] = float("{0:.2f}".format(ethShares / ( ethUptimeMin / 60 )))
 	else:
 		data['ethsharesperhour'] = 0
